Fix compile error due to wrong Eigen Vector type in SICP::point_to_plane()

+++ ICP.h
@@ -345,7 +345,7 @@ namespace SICP {
                     if(dual < par.stop) break;
                 }
                 /// C update (lagrange multipliers)
-                Eigen::VectorXf P = (Qn.array()*(X-Qp).array()).colwise().sum().transpose()-Z.array();
+                Eigen::VectorXd P = (Qn.array()*(X-Qp).array()).colwise().sum().transpose()-Z.array();
                 if(!par.use_penalty) C.noalias() += mu*P;
                 /// mu update (penalty)
                 if(mu < par.max_mu) mu *= par.alpha;
